fix(output): cache lipgloss renderers (#52)
* fix(output): cache lipgloss renderers

Cache and reuse lipgloss renderers to avoid concurrently querying the
terminal for background colors. Lipgloss HasDarkBackground function
queries the terminal for the background color. Creating multiple logger
instances with the same output will query the output multiple times
causing Log to hange and/or freeze.

To solve this, we store and retrieve the Lipgloss renderer in a global
sync.Map for each output along with caching the background/foreground
colors of the underlying Termenv Output.

Fixes: https://github.com/charmbracelet/log/issues/50

* Update logger.go

Co-authored-by: Carlos Alexandro Becker <caarlos0@users.noreply.github.com>

---------

Co-authored-by: Carlos Alexandro Becker <caarlos0@users.noreply.github.com>