fix(test): use sublogger generated from With() in race detection
The new sublogger generated from With() are never used, which may cause
optimization of the sublogger, and no race condition is detected.

Use a for loop to spawn a lots of goroutines to ensure the race
condition would happen in CI.

We also need to replace the bytes.Buffer to io.Discard, since the
bytes.Buffer is also not thread-safe.

Signed-off-by: yuguorui <yuguorui@pku.edu.cn>